mas_web/requirements.txt
streamlit==1.28.1
streamlit-echarts==0.4.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
